+++ src/pages/Auth/temp/hook/PC/index.module.scss
