+++ src/pages/Auth/temp/hook/mobile/VerifyCode.tsx
@@ -2,10 +2,10 @@ import { useState } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
 import cn from 'utils/ts/classNames';
 import { sendFindEmail, getAccount, findPassowrd } from 'api/user';
-import useInputCheck from 'pages/Auth/FindIdPassword/hook/useInputCheck';
-import { RegisterProp, CodeInfo } from 'pages/Auth/FindIdPassword/entity';
-import Modal from 'pages/Auth/FindIdPassword/mobile/Modal';
-import Input from 'pages/Auth/FindIdPassword/mobile/Input';
+import useInputCheck from 'pages/Auth/temp/hook/useInputCheck';
+import { RegisterProp, CodeInfo } from 'pages/Auth/temp/hook/entity';
+import Modal from 'pages/Auth/temp/hook/mobile/Modal';
+import Input from 'pages/Auth/temp/hook/mobile/Input';
 import style from './VerifyCode.module.scss';
 
 const CODE = ['first', 'second', 'third', 'fourth'] as const;
