+++ src/pages/Auth/temp/index.tsx
@@ -1,7 +1,7 @@
 import useMediaQuery from 'utils/hooks/useMediaQuery';
-import FindIdPasswordMobile from 'pages/Auth/FindIdPassword/mobile';
-import FindIdPasswordPC from 'pages/Auth/FindIdPassword/PC';
-import { FindProp } from 'pages/Auth/FindIdPassword/entity';
+import FindIdPasswordMobile from 'pages/Auth/temp/hook/mobile';
+import FindIdPasswordPC from 'pages/Auth/temp/hook/PC';
+import { FindProp } from 'pages/Auth/temp/hook/entity';
 
 export default function FindIdPassword({ type }: FindProp) {
   const { isMobile } = useMediaQuery();
